large changes - move formats to separate module section - add search commands for deezer, soundcloud, youtube - add translations (see modules/translations/lang) - remove uptime - misc. changes - run prettier

+++ modules/voice/ffmpeg.ts
@@ -80,8 +80,7 @@ export default class FFMpeg extends Transform {
   }
 
   public destroy(error?: Error) {
-    if (!this.instance.killed)
-      this.instance.kill();
+    if (!this.instance.killed) this.instance.kill();
     return super.destroy(error);
   }
 }
